+++ spec/helpers/application_helper_spec.rb
@@ -4,18 +4,22 @@ require 'rails_helper'
 
 describe ApplicationHelper, type: :helper do
   describe '#collection_title' do
+    it 'handles when the collection title does not include a druid' do
+      expect(helper.collection_title('A collection title')).to eq 'A collection title'
+    end
+
     it 'unmangles the collection title from the compound field' do
-      expect(helper.collection_title('foo-|-bar')).to eq 'foo'
+      expect(helper.collection_title('foo-|-bar')).to eq 'bar'
     end
   end
 
   describe '#collection_title_for_index_field' do
     it 'unmangles the collection title from the compound field' do
-      expect(helper.document_collection_title(value: 'foo-|-bar')).to eq 'foo'
+      expect(helper.document_collection_title(value: 'foo-|-bar')).to eq 'bar'
     end
 
     it 'handles multivalued fields' do
-      expect(helper.document_collection_title(value: ['foo-|-bar', 'baz-|-bop'])).to eq 'foo and baz'
+      expect(helper.document_collection_title(value: ['foo-|-bar', 'baz-|-bop'])).to eq 'bar and bop'
     end
   end